(BSR)[PRO] refactor: rename CollectiveOfferPreview screen file to fit its exported component naming
